Move completion_v2 into completion
