+++ compiler-lsp/analyzer/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod completion;
-pub mod completion_v2;
 pub mod definition;
 pub mod hover;
 pub mod locate;
